requirements.txt
Django==5.1.4
python-dotenv==1.0.1
requests==2.32.3
